requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
llama-index
llama-index-retrievers-bm25